r7415: * big change -- volker's new async winbindd from trunk

+++ source/sam/idmap_ldap.c
@@ -71,7 +71,8 @@ static NTSTATUS ldap_set_mapping(const DOM_SID *sid, unid_t id, int id_type)
 	else
 		fstrcpy( type, get_attr_key2string( sidmap_attr_list, LDAP_ATTR_GIDNUMBER ) );
 
-	pstr_sprintf(id_str, "%d", ((id_type & ID_USERID) ? id.uid : id.gid));	
+	pstr_sprintf(id_str, "%lu", ((id_type & ID_USERID) ? (unsigned long)id.uid :
+						 (unsigned long)id.gid));	
 	
 	smbldap_set_mod( &mods, LDAP_MOD_ADD, "objectClass", LDAP_OBJ_IDMAP_ENTRY );
 
@@ -500,9 +501,9 @@ static NTSTATUS ldap_get_sid_from_id(DOM_SID *sid, unid_t id, int id_type)
 		type = get_attr_key2string( idpool_attr_list, LDAP_ATTR_GIDNUMBER );
 
 	pstrcpy( suffix, lp_ldap_idmap_suffix() );
-	pstr_sprintf(filter, "(&(objectClass=%s)(%s=%d))",
+	pstr_sprintf(filter, "(&(objectClass=%s)(%s=%lu))",
 		LDAP_OBJ_IDMAP_ENTRY, type,  
-		((id_type & ID_USERID) ? id.uid : id.gid));
+		((id_type & ID_USERID) ? (unsigned long)id.uid : (unsigned long)id.gid));
 		
 	attr_list = get_attr_list( sidmap_attr_list );
 	rc = smbldap_search(ldap_state.smbldap_state, suffix, LDAP_SCOPE_SUBTREE, 
@@ -700,8 +701,8 @@ static NTSTATUS verify_idpool( void )
 			return NT_STATUS_UNSUCCESSFUL;
 		}
 		
-		fstr_sprintf( uid_str, "%d", luid );
-		fstr_sprintf( gid_str, "%d", lgid );
+		fstr_sprintf( uid_str, "%lu", (unsigned long)luid );
+		fstr_sprintf( gid_str, "%lu", (unsigned long)lgid );
 
 		smbldap_set_mod( &mods, LDAP_MOD_ADD, "objectClass", LDAP_OBJ_IDPOOL );
 		smbldap_set_mod( &mods, LDAP_MOD_ADD,
